Rename old test files which now contain the common test logic but do not run the actual tests.
