+++ clients/typescript/test/migrators/postgres/builder.test.ts
@@ -1,10 +1,6 @@
 import anyTest, { TestFn } from 'ava'
 import { makeMigration, parseMetadata } from '../../../src/migrators/builder'
-import {
-  ContextType,
-  bundleTests,
-  makeMigrationMetaData,
-} from '../builder.test'
+import { ContextType, bundleTests, makeMigrationMetaData } from '../builder'
 import { makePgDatabase } from '../../support/node-postgres'
 import { DatabaseAdapter } from '../../../src/drivers/node-postgres'
 import { PgBundleMigrator } from '../../../src/migrators'
